updated osinstaller

+++ dropscript.sh
@@ -3,51 +3,65 @@
 BINDIR="./binaries"
 FONTDIR="./fonts"
 
-echo "I should be run in the same directory as my configs. 
+# Display the prompt message
+echo "I should be run in the same directory as my configs.
 This script is intended to be a clone & run type script.
-Press <enter> to continue if you're sure..."
+Enter 'y' to continue."
 
-read 
+# Read user input
+read -r confirm
 
+# Check if the user confirmed with 'y' or 'Y'
+if [ "$confirm" != "y" ] && [ "$confirm" != "Y" ]; then
+    echo "Aborting... You did not confirm with 'y'."
+    exit 1
+fi
+
+# Continue with the rest of the script
+echo "Proceeding with the script..."
+
+# The rest of the script continues here...
+
+# Check if required directories exist
 if ! [ -d "dotConfigFiles" ] || ! [ -d "homeDotFiles" ]; then
-    echo "config directories not found, aborting.
-    expected to find './dotConfigfiles' & './homeDotFiles'"
-    exit 0
+    echo "Config directories not found, aborting.
+    Expected to find './dotConfigFiles' & './homeDotFiles'"
+    exit 1
 fi
 
 clear
 
 echo "Making config directories if needed..."
-if ! [ -d "$HOME"/.config ]; then 	mkdir --parents -v "$HOME"/.config; fi
-if ! [ -d "$HOME"/.local ]; then	mkdir --parents -v "$HOME"/.local/{bin,src,share}; fi
+# Create necessary directories if they don't exist
+mkdir -p "$HOME/.config" "$HOME/.local/bin" "$HOME/.local/src" "$HOME/.local/share"
 
 # Link all config files to $HOME/.config
-cd dotConfigFiles
-for fle in ./*
-do 	
-	fle=${fle##*/}
-	ln -sf "$PWD/$fle" "$HOME/.config/$fle"
+cd dotConfigFiles || exit
+for fle in ./*; do
+    fle=${fle##*/}
+    ln -sf "$PWD/$fle" "$HOME/.config/$fle"
 done
 
-cd ../homeDotFiles
-for fle in $(ls -lA | awk '{print $9}')
-do	
-	fle=${fle##*/}
-	ln -sf "$PWD/$fle" "$HOME/$fle" 
+cd ../homeDotFiles || exit
+for fle in /*; do
+    fle=${fle##*/}
+    ln -sf "$PWD/$fle" "$HOME/$fle"
 done
 
 cd ..
-mkdir "$HOME/.local/bin/cbin"
-# Link custom binaries to the users local bin path.
-ln -s "$BINDIR" "$HOME/.local/bin/cbin"
+
+# Create binary directory and link custom binaries
+mkdir -p "$HOME/.local/bin/cbin"
+ln -sf "$BINDIR" "$HOME/.local/bin/cbin"
 
 # Link fonts to the local fonts directory
 if ! [ -d "$HOME/.local/share/fonts" ]; then
-	ln -sr "$FONTDIR" "$HOME/.local/share/fonts"
+    ln -sr "$FONTDIR" "$HOME/.local/share/fonts"
 else
-	cp "$FONTDIR/**" "$HOME/.local/share/fonts/"
+    cp -r "$FONTDIR/"* "$HOME/.local/share/fonts/"
 fi
 fc-cache -fv "$HOME/.local/share/fonts"
 
-mkdir "$HOME"/{docs,proj,down};
-ln -s "$HOME"/down "$HOME"/Downloads
+# Create additional directories
+mkdir -p "$HOME/docs" "$HOME/proj" "$HOME/down"
+ln -s "$HOME/down" "$HOME/Downloads"
